+++ src/main/java/com/hipo/utill/AuthProcessor.java
@@ -1,4 +1,4 @@
-package com.hipo.domain.processor;
+package com.hipo.utill;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.AuthenticationManager;
